Single precision

+++ interface.go
@@ -46,7 +46,7 @@ type GameTreeNode interface {
 	GetChild(i int) GameTreeNode
 	// Get the probability of the ith child of this node.
 	// May only be called for nodes with Type == Chance.
-	GetChildProbability(i int) float64
+	GetChildProbability(i int) float32
 	// Sample one of the children of this node, according to the probability
 	// distribution. Only applicable for nodes with Type == Chance.
 	SampleChild() GameTreeNode
@@ -59,5 +59,5 @@ type GameTreeNode interface {
 	InfoSet(player int) InfoSet
 	// Utility returns this node's utility for the given player.
 	// It must only be called for nodes with type == Terminal.
-	Utility(player int) float64
+	Utility(player int) float32
 }
